Revert "refactor(test): tests/floodsub - addition of test_factory.rs"
This reverts commit 226ff569e7f8645a10581515426902103d4281d7.
