+++ floodsub/tests/floodsub/connections.rs
@@ -1,8 +1,32 @@
+use std::future::Future;
+use std::pin::Pin;
 use std::time::Duration;
-use testlib::any_memory_addr;
-use tokio::time::timeout;
 
-use crate::floodsub::flood_testlib::*;
+use libp2p::identity::{Keypair, PeerId};
+use libp2p::swarm::{Swarm, SwarmBuilder};
+use tokio::time::timeout;
+use tracing_futures::Instrument;
+
+use libp2p_pubsub_core::{Behaviour as PubsubBehaviour, Config};
+use libp2p_pubsub_floodsub::Protocol as Floodsub;
+use testlib::any_memory_addr;
+
+type Behaviour = PubsubBehaviour<Floodsub>;
+
+fn new_test_node(keypair: &Keypair, config: Config) -> Swarm<Behaviour> {
+    let peer_id = PeerId::from(keypair.public());
+    let transport = testlib::test_transport(keypair);
+    let behaviour = Behaviour::new(config, Default::default());
+    SwarmBuilder::with_executor(
+        transport,
+        behaviour,
+        peer_id,
+        |fut: Pin<Box<dyn Future<Output = ()> + Send>>| {
+            tokio::spawn(fut.in_current_span());
+        },
+    )
+    .build()
+}
 
 #[tokio::test]
 async fn connection_is_established() {
@@ -12,10 +36,10 @@ async fn connection_is_established() {
     let node_a_key = testlib::secp256k1_keypair(testlib::keys::TEST_KEYPAIR_A);
     let node_b_key = testlib::secp256k1_keypair(testlib::keys::TEST_KEYPAIR_B);
 
-    let mut node_a = new_test_node(&node_a_key);
+    let mut node_a = new_test_node(&node_a_key, Default::default());
     testlib::swarm::should_listen_on_address(&mut node_a, any_memory_addr());
 
-    let mut node_b = new_test_node(&node_b_key);
+    let mut node_b = new_test_node(&node_b_key, Default::default());
     testlib::swarm::should_listen_on_address(&mut node_b, any_memory_addr());
 
     let (node_a_addr, _node_b_addr) = timeout(
