--- floodsub/tests/floodsub/flood_testlib/mod.rs
@@ -1,3 +1,0 @@
-pub use test_factory::*;
-
-mod test_factory;
