+++ floodsub/tests/floodsub/mod.rs
@@ -1,4 +1,3 @@
 mod connections;
-mod flood_testlib;
 mod routing;
 mod subscriptions;
